+++ unikmer/cmd/diff.go
@@ -42,7 +42,7 @@ var diffCmd = &cobra.Command{
 
 Attentions:
   0. The first file should be sorted.
-  1. The 'canonical' flags of all files should be consistent.
+  1. The 'canonical/scaled/hashed' flags of all files should be consistent.
   2. By default taxids in the 2nd and later files are ignored.
   3. You can switch on flag -t/--compare-taxid , and input
      files should ALL have or don't have taxid information.