+++ unikmer/cmd/sample.go
@@ -40,7 +40,7 @@ var sampleCmd = &cobra.Command{
 The Sampling type is fixed sampling.
 
 Attentions:
-  1. The 'canonical' flags of all files should be consistent.
+  1. The 'canonical/scaled/hashed' flags of all files should be consistent.
   2. Input files should ALL have or don't have taxid information.
 
 `,