+++ unikmer/cmd/sort.go
@@ -43,7 +43,7 @@ var sortCmd = &cobra.Command{
 	Long: `Sort k-mers in binary files to reduce file size
 
 Attentions:
-  1. The 'canonical' flags of all files should be consistent.
+  1. The 'canonical/scaled/hashed' flags of all files should be consistent.
   2. Input files should ALL have or don't have taxid information.
 
 Notes: