+++ unikmer/cmd/split.go
@@ -44,7 +44,7 @@ var splitCmd = &cobra.Command{
 	Long: `Split k-mers into sorted chunk files
 
 Attentions:
-  1. The 'canonical' flags of all files should be consistent.
+  1. The 'canonical/scaled/hashed' flags of all files should be consistent.
   2. Input files should ALL have or don't have taxid information.
   
 Tips: