+++ unikmer/cmd/view.go
@@ -40,7 +40,7 @@ var viewCmd = &cobra.Command{
 	Long: `Read and output binary format to plain text
 
 Attentions:
-  1. The 'canonical' flags of all files should be consistent.
+  1. The 'canonical/scaled/hashed' flags of all files should be consistent.
   2. Input files should ALL have or don't have taxid information.
   
 `,